🎨[#1929] Fix test structure
 - Add explicit assertion
 - Move helper function from formio package
