+++ src/openforms/api/tests/test_validators.py
@@ -1,7 +1,5 @@
-from unittest.mock import patch
-
 from django.core.files.uploadedfile import SimpleUploadedFile
-from django.test import SimpleTestCase as TestCase
+from django.test import SimpleTestCase, override_settings
 
 from rest_framework.exceptions import ValidationError
 
@@ -11,10 +9,8 @@ PIXEL_GIF = b"GIF89a\x01\x00\x01\x00\x81\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00
 PIXEL_PNG = b"\x89PNG\r\n\x1a\n\x00\x00\x00\rIHDR\x00\x00\x00\x01\x00\x00\x00\x01\x08\x02\x00\x00\x00\x90wS\xde\x00\x00\x00\x0cIDATx\x9cc```\x00\x00\x00\x04\x00\x01\xf6\x178U\x00\x00\x00\x00IEND\xaeB`\x82"
 
 
-@patch(
-    "openforms.api.validators._", lambda s: s
-)  # decouple from translations by returning msgid
-class MimeTypeValidatorTests(TestCase):
+@override_settings(LANGUAGE_CODE="en")
+class MimeTypeValidatorTests(SimpleTestCase):
     CORRECT_GIF = SimpleUploadedFile("pixel.gif", PIXEL_GIF, content_type="image/gif")
 
     def setUp(self):
@@ -22,14 +18,21 @@ class MimeTypeValidatorTests(TestCase):
         self.CORRECT_GIF.seek(0)
 
     def test_accepts_correct_mime_types(self):
-        validators.MimeTypeValidator()(self.CORRECT_GIF)
+        validator = validators.MimeTypeValidator()
+
+        try:
+            validator(self.CORRECT_GIF)
+        except ValidationError as e:
+            self.fail(f"Correct file failed validation: {e}")
 
     def test_content_inconsistent_with_mime_type(self):
         file = SimpleUploadedFile("pixel.png", PIXEL_GIF, content_type="image/png")
+        validator = validators.MimeTypeValidator()
+
         with self.assertRaisesMessage(
             ValidationError, "The file 'pixel.png' is not a .png."
         ):
-            validators.MimeTypeValidator()(file)
+            validator(file)
 
     def test_fallback_to_extension(self):
         file = SimpleUploadedFile(
@@ -37,40 +40,52 @@ class MimeTypeValidatorTests(TestCase):
             PIXEL_PNG,
             content_type="application/octet-stream",  # Maybe client doesn't know
         )
+        validator = validators.MimeTypeValidator()
+
         with self.assertRaisesMessage(
             ValidationError, "The file 'pixel.jpg' is not a .jpg."
         ):
-            validators.MimeTypeValidator()(file)
+            validator(file)
 
     def test_accepts_unknown_extensions(self):
         file = SimpleUploadedFile(
             "pixel.gif", PIXEL_GIF, content_type="application/octet-stream"
         )
+        validator = validators.MimeTypeValidator()
+
         try:
-            validators.MimeTypeValidator()(file)
+            validator(file)
         except ValidationError as e:
             self.fail(f"Valid file failed validation: {e}")
 
     def test_star_wildcard_in_allowed_mimetypes(self):
+        validator = validators.MimeTypeValidator({"*"})
+
         try:
-            validators.MimeTypeValidator({"*"})(self.CORRECT_GIF)
+            validator(self.CORRECT_GIF)
         except ValidationError as e:
             self.fail(f"Valid file failed validation: {e}")
 
     def test_empty_allowed_mimetypes(self):
+        validator = validators.MimeTypeValidator({})
+
         try:
-            validators.MimeTypeValidator({})(self.CORRECT_GIF)
+            validator(self.CORRECT_GIF)
         except ValidationError as e:
             self.fail(f"Valid file failed validation: {e}")
 
     def test_mime_type_in_allowed_mimetypes(self):
+        validator = validators.MimeTypeValidator({"image/gif"})
+
         try:
-            validators.MimeTypeValidator({"image/gif"})(self.CORRECT_GIF)
+            validator(self.CORRECT_GIF)
         except ValidationError as e:
             self.fail(f"Valid file failed validation: {e}")
 
     def test_mime_type_not_allowed(self):
+        validator = validators.MimeTypeValidator({"image/png"})
+
         with self.assertRaisesMessage(
             ValidationError, "The file 'pixel.gif' is not a valid file type."
         ):
-            validators.MimeTypeValidator({"image/png"})(self.CORRECT_GIF)
+            validator(self.CORRECT_GIF)
