+++ src/openforms/api/validators.py
@@ -7,7 +7,20 @@ from django.utils.translation import gettext_lazy as _
 import magic
 from rest_framework import serializers
 
-from openforms.formio.service import mimetype_allowed
+
+def mimetype_allowed(mime_type: str, allowed_mime_types: Container[str]) -> bool:
+    """
+    Test if the file mime type passes the allowed_mime_types Formio configuration.
+    """
+    #  no allowlist specified -> everything is allowed
+    if not allowed_mime_types:
+        return True
+
+    # wildcard specified -> everything is allowed
+    if "*" in allowed_mime_types:
+        return True
+
+    return mime_type in allowed_mime_types
 
 
 class AllOrNoneRequiredFieldsValidator:
